We place the deadline as OP_VOTE_DEADLINE within the proposal first public output script. votes will not be accepted unless placed in a block before the deadline.
We get rid of redeemed proposals. Instead, we search for alarms that go
off after a certain number of blocks.

Proposals will now be stored as <deadline,txnhash> -> total_votes

This way we can create a cursor and go through proposals to find the
ones that:

 * passed, and
 * that should be displayed in the gui (only proposals that have
    >10% vote are displayed in gui screens)

Stakes no longer replace votes directly. Votes are kept separate (the
purpose is to make the blockchain less confusing in the end)

The total number of active voters is now calculated correctly. This is
based on when the user last voted, regardless of when the proposal deadline
occurs

Removed preamble stuff. First, it's not needed, and second, a vote script
now always has a standard size of <txnhash size> + 1 byte for OP_VOTE
(we don't sign voting inputs with a private key